# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Movena replacement core contracts",
    "version": "2.0.0-alpha.1",
    "description": "Typed initial recovery slice. Other compatibility endpoints remain inventoried in baseline.json."
  },
  "paths": {
...
    "/api/v2/auth/me": {
      "get": {
        "operationId": "getCurrentUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
    "/api/v2/auth/login": {
      "post": {
        "operationId": "login",
        "security": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginRequest"
              }
            }
          }
        }
      }
    },
    "/api/v2/auth/register": {
      "post": {
        "operationId": "register",
        "security": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummary"
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterRequest"
              }
            }
          }
        }
      }
    },
    "/api/v2/auth/forgot-password": {
      "post": {
        "operationId": "forgotPassword",
        "security": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRequest"
              }
            }
          }
        }
      }
    },
    "/api/v2/auth/reset-password": {
      "post": {
        "operationId": "resetPassword",
        "security": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          }
        }
      }
    },
    "/api/v2/auth/verify-email": {
      "post": {
        "operationId": "verifyEmail",
        "security": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          }
        }
      }
    },
...
    "/api/v2/patient/today": {
      "get": {
        "operationId": "getToday",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientTodayResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
...
    "/api/v2/patient/consents": {
      "get": {
        "operationId": "getConsents",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsentRecord"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
...
    "/api/v2/patient/notifications": {
      "get": {
        "operationId": "getNotifications",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
...
    "/api/v2/patient/reports": {
      "get": {
        "operationId": "getPatientReports",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgressReportSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
...
    "/api/v2/patient/adherence": {
      "post": {
        "operationId": "saveCheckIn",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdherenceDetail"
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdherenceCreate"
              }
            }
          }
        }
      }
    },
...
    "/api/v2/patient/appointments": {
      "get": {
        "operationId": "getPatientAppointments",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
...
    "/api/v2/scheduling/appointments": {
      "post": {
        "operationId": "bookAppointment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentSummary"
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentCreate"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthMessageResponse": {
        "properties": {
          "status": {
            "default": "success",
            "title": "Status",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "AuthMessageResponse",
        "type": "object"
      },
...
      "UserLoginRequest": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "UserLoginRequest",
        "type": "object"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "user": {
            "$ref": "#/components/schemas/UserSummary"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}